Moved the Header to the App component
