--- src/components/Header/Header.module.css
@@ -1,17 +1,0 @@
-.Header {
-  min-height: fit-content;
-  position: sticky;
-  top: 0;
-  left: 0;
-  z-index: 10;
-  background-color: var(--background-color);
-  transition: background-color 0.2s ease-in-out;
-}
-
-nav {
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-  padding: 1rem 2rem;
-  box-shadow: 0 2px 4px rgba(0, 0, 0, 0.1);
-}
